rename repository.common package to repository.base
