+++ src/main/java/org/sbml4j/repository/base/GraphBaseEntityRepository.java
@@ -11,7 +11,7 @@
  * 
  * For a full list of authors, please refer to the file AUTHORS.
  */
-package org.sbml4j.repository.common;
+package org.sbml4j.repository.base;
 
 import org.sbml4j.model.base.GraphBaseEntity;
 import org.springframework.data.neo4j.repository.Neo4jRepository;
