+++ src/main/java/org/sbml4j/service/GraphBaseEntityService.java
@@ -18,7 +18,7 @@ import java.util.UUID;
 import org.sbml4j.config.SBML4jConfig;
 import org.sbml4j.model.base.GraphBaseEntity;
 import org.sbml4j.model.base.GraphEnum.AnnotationName;
-import org.sbml4j.repository.common.GraphBaseEntityRepository;
+import org.sbml4j.repository.base.GraphBaseEntityRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
